pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ddx"
version = "0.1.0"
description = "Document map→reduce extractor for technical due diligence"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.30.0",
  "python-dotenv>=1.0.1",
  "pypdf2>=3.0.0",
  "pdfminer.six>=20231228",
  "pillow>=10.0.0",
  "pytesseract>=0.3.10",
  "pdf2image>=1.17.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
